+++ commerce/payment/src/main/java/ru/yandex/practicum/payment/service/PaymentService.java
@@ -3,16 +3,17 @@ package ru.yandex.practicum.payment.service;
 import ru.yandex.practicum.dto.order.OrderDto;
 import ru.yandex.practicum.dto.payment.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
     PaymentDto createPayment(OrderDto orderDto);
 
-    double calculateTotalCost(OrderDto orderDto);
+    BigDecimal calculateTotalCost(OrderDto orderDto);
 
     void refundPayment(UUID orderId);
 
-    double calculateProductCost(OrderDto orderDto);
+    BigDecimal calculateProductCost(OrderDto orderDto);
 
     void failedPayment(UUID orderId);
 }
